+++ spec/warning_signs/unhandled_deprecation_error_spec.rb
@@ -1,0 +1,8 @@
+module WarningSigns
+  RSpec.describe UnhandledDeprecationError do
+    it "handles warnings" do
+      exe = UnhandledDeprecationError.new("Feature x is going away")
+      expect(exe.message.ends_with?("exceptions list in .warning_signs.yml.")).to be_truthy
+    end
+  end
+end